Don't use str for inner classes in nested JSONs
Python supports inner classes and they work just fine in incoming, it's just a
matter of scoping them correctly :) Also, the documentation shouldn't hint
users to use strings instead because it's not as powerful and it can be
harmful.

For example, using proper classnames, Python will complain at compile time if a
name is wrong. If using strings, the error would go unnoticed until the code
path is executed.